added task

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -11,7 +11,7 @@
 
 int main(void)
 {
-	long int i, j, mx = 0, n = 612852475143;
+	long long int i, j, mx = 0, n = 612852475143;
 
 	for (i = 2; i * i <= n; i++)
 	{
